requirements.txt
anyio==4.10.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudevents==1.11.0
colorama==0.4.6
deprecation==2.1.0
exceptiongroup==1.3.0
Flask==3.1.2
functions-framework==3.9.2
google-api-core==2.25.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-cloud-bigquery==3.36.0
google-cloud-core==2.4.3
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.3.1
packaging==25.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
tomli==2.2.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
uvicorn-worker==0.3.0
watchdog==6.0.0
Werkzeug==3.1.3
